Move Healtchecks Reducers
