+++ plugins/services/src/js/reducers/JSONSingleContainerParser.js
@@ -3,7 +3,7 @@ import {JSONParser as constraints} from './serviceForm/JSONReducers/Constraints'
 import {JSONParser as fetch} from './serviceForm/JSONReducers/Artifacts';
 import {JSONParser as environmentVariables} from './serviceForm/EnvironmentVariables';
 import {JSONParser as externalVolumes} from './serviceForm/JSONReducers/ExternalVolumes';
-import {JSONParser as healthChecks} from './serviceForm/HealthChecks';
+import {JSONParser as healthChecks} from './serviceForm/JSONReducers/HealthChecks';
 import {JSONParser as labels} from './serviceForm/JSONReducers/Labels';
 import {JSONParser as localVolumes} from './serviceForm/LocalVolumes';
 import {JSONParser as portDefinitions} from './serviceForm/PortDefinitions';
@@ -41,4 +41,3 @@ module.exports = [
   residency,
   fetch
 ];
-
