+++ plugins/services/src/js/reducers/JSONSingleContainerReducers.js
@@ -2,7 +2,7 @@ import {JSONReducer as constraints} from './serviceForm/JSONReducers/Constraints
 import {JSONReducer as container} from './serviceForm/Container';
 import {JSONReducer as env} from './serviceForm/EnvironmentVariables';
 import {JSONReducer as fetch} from './serviceForm/JSONReducers/Artifacts';
-import {JSONReducer as healthChecks} from './serviceForm/HealthChecks';
+import {JSONReducer as healthChecks} from './serviceForm/JSONReducers/HealthChecks';
 import {JSONReducer as labels} from './serviceForm/JSONReducers/Labels';
 import {JSONReducer as portDefinitions} from './serviceForm/PortDefinitions';
 import {JSONReducer as residency} from './serviceForm/Residency';
